refactor: move metrics tracker into shared utils
