+++ src/plugin/src/reporting/metrics.js
@@ -1,1 +1,1 @@
-export { createMetricsTracker } from "./metrics-tracker.js";
+export { createMetricsTracker } from "../../../shared/metrics-utils.js";
